test(sdk): Add integration tests for remuneration and withdrawal
