+++ sdk/src/client/recipient.rs
@@ -63,7 +63,6 @@ impl RecipientClient {
             .getPaymentStatus(tab_id)
             .call()
             .await
-            .map_err(alloy::contract::Error::from)
             .map_err(TabPaymentStatusError::from)?;
 
         Ok(TabPaymentStatus {
