+++ sdk/tests/common/mod.rs
@@ -1,0 +1,7 @@
+use std::time::Duration;
+
+pub fn get_now() -> Duration {
+    std::time::SystemTime::now()
+        .duration_since(std::time::UNIX_EPOCH)
+        .unwrap()
+}
